requirements.txt
streamlit 
pillow
google.generativeai
datetime
